添加模拟器包
1、修复分段同步逻辑；
2、修复获取手环信息绑定状态存储逻辑；
3、demo更新gps119固件包；
4、添加蓝牙写完数据的回调方法；

+++ IDOBlue/IDOBlue/Models/ViewModels/SetViewModel/AllAlarmViewModel.m
@@ -86,7 +86,7 @@
         };
         newAlarmModel.addAlarmComplete = addAlarmComplete;
         newFuncVc.model = newAlarmModel;
-        newFuncVc.title = [NSString stringWithFormat:@"编辑闹钟(%ld)",alarmModel.alarmId];
+        newFuncVc.title = [NSString stringWithFormat:@"编辑闹钟(%ld)",(long)alarmModel.alarmId];
         [viewController.navigationController pushViewController:newFuncVc animated:YES];
     };
 }
@@ -116,7 +116,8 @@
         if(alarmModel.isOpen && alarmModel.isSync) {
             LabelCellModel * model = [[LabelCellModel alloc]init];
             model.typeStr = @"oneLabel";
-            NSString * alarmStr = [NSString stringWithFormat:@"闹钟 : %ld \n时间 : %02ld:%02ld \n重复 : %@",alarmModel.alarmId,alarmModel.hour,alarmModel.minute,
+            NSString * alarmStr = [NSString stringWithFormat:@"闹钟 : %ld \n时间 : %02ld:%02ld \n重复 : %@",(long)alarmModel.alarmId,
+                                   (long)alarmModel.hour,(long)alarmModel.minute,
                                    [self weeksStrWithRepeats:alarmModel.repeat]];
             model.data = @[alarmStr];
             model.cellHeight = 80.0f;
